# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "StellarKraal API",
    "version": "1.0.0",
    "description": "Livestock-backed micro-lending API on Stellar/Soroban. Write endpoints return an unsigned XDR transaction the client must sign with Freighter and submit to the Stellar network."
  },
  "servers": [
    { "url": "/api/v1", "description": "Versioned API" },
    { "url": "/api", "description": "Unversioned (deprecated, redirects to /api/v1)" }
  ],
...
  "security": [
    { "bearerAuth": [] }
  ],
  "paths": {
...
    "/collateral/register": {
      "post": {
        "tags": ["collateral"],
        "summary": "Register livestock collateral (on-chain)",
        "description": "Builds an unsigned Soroban transaction to register livestock as on-chain collateral.",
        "operationId": "registerCollateral",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RegisterCollateralRequest" } } } },
        "responses": {
          "200": { "description": "Unsigned XDR transaction", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/XdrResponse" } } } },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
    "/collateral/{id}": {
      "patch": {
        "tags": ["collateral"],
        "summary": "Partially update collateral fields",
        "operationId": "updateCollateral",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/V1UpdateCollateralRequest" } } }
        },
        "responses": {
          "200": { "description": "Updated record", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CollateralRecord" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "422": { "$ref": "#/components/responses/ValidationError" }
        }
      },
      "delete": {
        "tags": ["collateral"],
        "summary": "Soft-delete a collateral record",
        "operationId": "deleteCollateral",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Deleted", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DeletedResponse" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
    "/loan/request": {
      "post": {
        "tags": ["loan"],
        "summary": "Request a new loan",
        "operationId": "requestLoan",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LoanRequestBody" } } } },
        "responses": {
          "200": { "description": "Unsigned XDR transaction", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/XdrResponse" } } } },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/loan/repay": {
      "post": {
        "tags": ["loan"],
        "summary": "Repay a loan",
        "description": "Requires `Idempotency-Key` header to prevent duplicate submissions.",
        "operationId": "repayLoan",
        "parameters": [{ "name": "Idempotency-Key", "in": "header", "required": true, "schema": { "type": "string" }, "description": "Client-generated unique key for idempotent replay" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LoanRepayBody" } } } },
        "responses": {
          "200": {
            "description": "Unsigned XDR transaction",
            "headers": {
              "X-Idempotent-Replayed": {
                "schema": { "type": "string", "enum": ["true"] },
                "description": "Present with value `true` when the response was served from the idempotency cache"
              }
            },
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/XdrResponse" } } }
          },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/loan/liquidate": {
      "post": {
        "tags": ["loan"],
        "summary": "Liquidate an undercollateralised loan",
        "operationId": "liquidateLoan",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LoanLiquidateBody" } } } },
        "responses": {
          "200": { "description": "Unsigned XDR transaction", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/XdrResponse" } } } },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
    "/loan/{id}": {
      "get": {
        "tags": ["loan"],
        "summary": "Fetch a loan record",
        "operationId": "getLoan",
        "security": [],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "minimum": 0 } }],
        "responses": {
          "200": { "description": "Loan record", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ScValResponse" } } } },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "delete": {
        "tags": ["loan"],
        "summary": "Soft-delete a loan record",
        "operationId": "deleteLoan",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Deleted", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DeletedResponse" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/loans": {
      "get": {
        "tags": ["loan"],
        "summary": "List loans (paginated)",
        "operationId": "listLoans",
        "security": [],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "minimum": 1, "default": 1 } },
          { "name": "pageSize", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 20 } }
        ],
        "responses": {
          "200": { "description": "Paginated loan list", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LoanPage" } } } },
          "400": { "$ref": "#/components/responses/ValidationError" }
        }
      }
    },
...
    "/oracle/price-update": {
      "post": {
        "tags": ["oracle"],
        "summary": "Invalidate appraisal cache",
        "operationId": "oraclePriceUpdate",
        "responses": {
          "200": { "description": "Cache invalidated", "content": { "application/json": { "schema": { "type": "object", "properties": { "invalidated": { "type": "boolean", "example": true } } } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/webhooks": {
      "post": {
        "tags": ["webhooks"],
        "summary": "Register a webhook",
        "operationId": "registerWebhook",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object", "required": ["url"], "properties": { "url": { "type": "string", "format": "uri" } } } } }
        },
        "responses": {
          "201": { "description": "Registered", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/WebhookRecord" } } } },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/alerts/webhook": {
      "post": {
        "tags": ["admin"],
        "summary": "AWS SNS alert receiver",
        "description": "Receives AWS SNS notifications (SubscriptionConfirmation and Notification). Used for backup failure alerts.",
        "operationId": "alertsWebhook",
        "security": [],
        "parameters": [
          { "name": "x-amz-sns-message-type", "in": "header", "schema": { "type": "string", "enum": ["SubscriptionConfirmation", "Notification"] } }
        ],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object" } } } },
        "responses": {
          "200": { "description": "Accepted" }
        }
      }
    },
    "/admin/webhooks": {
      "get": {
        "tags": ["admin"],
        "summary": "List registered webhooks",
        "operationId": "listWebhooks",
        "responses": {
          "200": { "description": "Webhook list", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/WebhookRecord" } } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/admin/webhooks/logs": {
      "get": {
        "tags": ["admin"],
        "summary": "List webhook delivery logs",
        "operationId": "listWebhookLogs",
        "responses": {
          "200": { "description": "Delivery logs", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/WebhookLog" } } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
    "/settings/{wallet}": {
      "get": {
        "tags": ["settings"],
        "summary": "Get user settings",
        "operationId": "getSettings",
        "security": [],
        "parameters": [{ "name": "wallet", "in": "path", "required": true, "schema": { "type": "string" }, "description": "Stellar public key" }],
        "responses": {
          "200": { "description": "User settings", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserSettings" } } } }
        }
      },
      "put": {
        "tags": ["settings"],
        "summary": "Update user settings",
        "operationId": "updateSettings",
        "parameters": [{ "name": "wallet", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserSettingsUpdate" } } } },
        "responses": {
          "200": { "description": "Updated settings", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserSettings" } } } },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "HS256 JWT issued by POST /api/auth/login. Required on all POST, PUT, DELETE, and PATCH requests except /api/auth/* endpoints. Pass as: Authorization: Bearer <token>"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": { "type": "string" },
          "code": { "type": "string" },
          "details": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": { "type": "string" }
            }
          }
        }
      },
...
      "RegisterCollateralRequest": {
        "type": "object",
        "required": ["owner", "animal_type", "count", "appraised_value"],
        "properties": {
          "owner": { "type": "string", "description": "Stellar public key (G...)", "example": "GAAZI4TCR3TY5OJHCTJC2A4QSY6CJWJH5IAJTGKIN2ER7LBNVKOCCWN" },
          "animal_type": { "type": "string", "example": "cattle" },
          "count": { "type": "integer", "minimum": 1, "example": 5 },
          "appraised_value": { "type": "integer", "minimum": 1, "description": "Total appraised value in token base units", "example": 5000000 }
        }
      },
...
      "V1UpdateCollateralRequest": {
        "type": "object",
        "properties": {
          "animal_type": { "type": "string", "minLength": 1, "example": "goat" },
          "count": { "type": "integer", "minimum": 1, "example": 3 },
          "appraised_value": { "type": "number", "exclusiveMinimum": 0, "example": 3500000 }
        },
        "additionalProperties": false,
        "minProperties": 1
      },
      "CollateralRecord": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "owner": { "type": "string" },
          "animal_type": { "type": "string" },
          "count": { "type": "integer" },
          "appraised_value": { "type": "integer" },
          "deleted_at": { "type": "string", "format": "date-time", "nullable": true }
        }
      },
...
      "LoanRequestBody": {
        "type": "object",
        "required": ["borrower", "collateral_id", "amount"],
        "properties": {
          "borrower": { "type": "string", "description": "Stellar public key", "example": "GAAZI4TCR3TY5OJHCTJC2A4QSY6CJWJH5IAJTGKIN2ER7LBNVKOCCWN" },
          "collateral_id": { "type": "integer", "minimum": 0, "example": 1 },
          "amount": { "type": "integer", "minimum": 1, "description": "Gross loan amount in token base units", "example": 3000000 }
        }
      },
      "LoanRepayBody": {
        "type": "object",
        "required": ["borrower", "loan_id", "amount"],
        "properties": {
          "borrower": { "type": "string", "example": "GAAZI4TCR3TY5OJHCTJC2A4QSY6CJWJH5IAJTGKIN2ER7LBNVKOCCWN" },
          "loan_id": { "type": "integer", "minimum": 0, "example": 1 },
          "amount": { "type": "integer", "minimum": 1, "example": 1000000 }
        }
      },
      "LoanLiquidateBody": {
        "type": "object",
        "required": ["liquidator", "loan_id", "repay_amount"],
        "properties": {
          "liquidator": { "type": "string", "example": "GAAZI4TCR3TY5OJHCTJC2A4QSY6CJWJH5IAJTGKIN2ER7LBNVKOCCWN" },
          "loan_id": { "type": "integer", "minimum": 0, "example": 1 },
          "repay_amount": { "type": "integer", "minimum": 1, "example": 500000 }
        }
      },
...
      "LoanPage": {
        "type": "object",
        "properties": {
          "data": { "type": "array", "items": { "type": "object" } },
          "total": { "type": "integer" },
          "page": { "type": "integer" },
          "pageSize": { "type": "integer" }
        }
      },
...
      "XdrResponse": {
        "type": "object",
        "properties": {
          "api_version": { "type": "string", "example": "v1" },
          "xdr": { "type": "string", "description": "Base64-encoded unsigned Soroban transaction XDR" }
        }
      },
      "ScValResponse": {
        "type": "object",
        "properties": {
          "api_version": { "type": "string", "example": "v1" },
          "result": { "description": "Raw Soroban ScVal result" }
        }
      },
...
      "WebhookRecord": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "url": { "type": "string", "format": "uri" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "WebhookLog": {
        "type": "object",
        "properties": {
          "webhookId": { "type": "string" },
          "event": { "type": "string" },
          "status": { "type": "integer" },
          "timestamp": { "type": "string", "format": "date-time" }
        }
      },
...
      "UserSettings": {
        "type": "object",
        "properties": {
          "walletAddress": { "type": "string" },
          "joinDate": { "type": "string", "format": "date-time" },
          "notifications": {
            "type": "object",
            "properties": {
              "loanApproved": { "type": "boolean" },
              "loanRepaid": { "type": "boolean" },
              "liquidationWarning": { "type": "boolean" }
            }
          },
          "language": { "type": "string", "example": "en" },
          "currency": { "type": "string", "example": "USD" }
        }
      },
      "UserSettingsUpdate": {
        "type": "object",
        "properties": {
          "notifications": {
            "type": "object",
            "properties": {
              "loanApproved": { "type": "boolean" },
              "loanRepaid": { "type": "boolean" },
              "liquidationWarning": { "type": "boolean" }
            }
          },
          "language": { "type": "string", "minLength": 2, "maxLength": 10 },
          "currency": { "type": "string", "minLength": 3, "maxLength": 6 }
        }
      },
      "DeletedResponse": {
        "type": "object",
        "properties": {
          "deleted": { "type": "boolean", "example": true },
          "id": { "type": "string" }
        }
      },
...
    },
    "responses": {
      "ValidationError": {
        "description": "Request validation failed",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
      },
      "Unauthorized": {
        "description": "Missing or invalid JWT",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
      },
      "RateLimited": {
        "description": "Too many requests",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
      },
      "InternalError": {
        "description": "Internal server error",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
      }
    }
  }
}